move mutations to local file, put all in subdir activity/
