+++ api-lib/gql/mutations.ts
@@ -534,15 +534,3 @@ export async function insertInteractionEvents(
   );
   return;
 }
-
-export async function insertActivity(
-  activity: GraphQLTypes['activities_insert_input']
-) {
-  await adminClient.mutate(
-    {
-      insert_activities_one: [{ object: activity }, { __typename: true }],
-    },
-    { operationName: 'insertActivity' }
-  );
-  return;
-}
